feat: MRKT-14 Move hoc to components
