+++ apps/marketplace/src/app/layout.tsx
@@ -5,10 +5,10 @@ import { AppRouterCacheProvider } from "@mui/material-nextjs/v13-appRouter";
 import { FunctionComponent, ReactNode } from "react";
 import theme from "@newm-web/theme";
 import { Provider } from "react-redux";
+import { StyledComponentsRegistry } from "../components";
 import Header from "../components/header/Header";
 import store from "../store";
 import Footer from "../components/footer/Footer";
-import { StyledComponentsRegistry } from "../common";
 
 interface RootLayoutProps {
   readonly children: ReactNode;
