+++ apps/marketplace/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./header";
 export { Header } from "./header";
 export * from "./footer/Footer";
 export * from "./Songs";
+export * from "./StyledComponentsRegistry";
